+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 <hr>
 
+## [1.3.17-RC1] 2023-07-31
+
+### Added
+
+- API Core - Operator bulk force success below orders with cod collected {string}:
+- API Core - Operator force success waypoint with cod collected as {string} using route manifest:
+- API Route - Operator run FM PAJ auto route cron job for date {string}
+
+<hr>
+
 ## [1.3.16] 2023-07-28
 
 ### Added